mail-react: Rename the responsive CSS module to `responsiveCss.ts` (#6139)
A file should not be named after one of its several exports.
