+++ packages/mail-react/src/components/divider/generateResponsiveDividerCss.ts
@@ -1,4 +1,4 @@
-import { generateResponsiveVariantCss, type ResponsiveVariantProperties } from "../../styles/generateResponsiveVariantCss.js";
+import { generateResponsiveVariantCss, type ResponsiveVariantProperties } from "../../styles/responsiveCss.js";
 import type { DividerVariantStyles, Theme } from "../../theme/themeTypes.js";
 
 const dividerProperties: ResponsiveVariantProperties<DividerVariantStyles> = [
